+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Whenever `minor` is bumped in the `major.minor.patch` version (e.g. `1.0.2 -> 1.
 - v1.1.6 (unreleased)
   - Ship `web/` in the package so that `bake-existing` can run
   - Move campaign annotations storage from `data/outputs/` to `data/annotations/` with automatic migration
+  - Make dynamic assignment optimize allocation with best effort, especially during coldstart
 - v1.1.5
   - Add `show_progress` flag to dashboard
   - Default to a single user for single-stream and dynamic campaigns if not specified